refactor: extract camera-ready run_campaign orchestrator (#3385)
Move the remaining ~1170-line `run_campaign` orchestrator out of the
`camera_ready_campaign.py` facade into a new
`robot_sf/benchmark/camera_ready/campaign.py` module, completing the #3385
package decomposition. The legacy module stays a thin compatibility facade.

`run_campaign` now accepts its filesystem-/subprocess-touching collaborators
(`prepare_campaign_preflight`, `run_batch`, `compute_aggregates_with_ci`,
`export_publication_bundle`) as optional injected callables, mirroring the
pattern already used by `camera_ready/_preflight.prepare_campaign_preflight`.
The facade `run_campaign` wrapper injects its own module-level bindings so
existing tests that monkeypatch `camera_ready_campaign.<name>` keep working
unchanged.

No behavior change: the orchestration body is moved verbatim; only the
dependency-resolution scaffolding is added.

Validation:
- uv run pytest tests/benchmark/test_camera_ready_campaign.py -> 127 passed
- downstream importer tests (run_camera_ready_benchmark, run_benchmark_release,
  amv_actuation_sensitivity_sweep, route_clearance_fail_closed,
  predictive_success_campaign_tags) -> 43 passed
- ruff check + format on camera_ready_campaign.py and camera_ready/ -> clean
